@@ -1,5 +1,3 @@
-from typing import Dict
-
 import xmltodict
 from pyinaturalist import Observation
 
@@ -37,7 +35,7 @@ def test_to_gpx__to_file(tmp_path):
     _validate_point(points[0])
 
 
-def _validate_point(point: Dict):
+def _validate_point(point: dict):
     assert point['@lat'] == '32.8430971478' and point['@lon'] == '-117.2815829044'
     assert point['time'] == '2020-05-09T06:01:00-07:00'
     assert 'Dirona picta' in point['cmt']
